+++ robotframework_reportportal/variables.py
@@ -13,18 +13,26 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License
 
+import sys
+
 from distutils.util import strtobool
 from os import path
+from typing import Optional, Union, Dict, Any, TextIO
 from warnings import warn
 
 from reportportal_client.logs.log_manager import MAX_LOG_BATCH_PAYLOAD_SIZE
 from robot.libraries.BuiltIn import BuiltIn, RobotNotRunningError
 
 # This is a storage for the result visitor
-_variables = {}
+_variables: Dict[str, Any] = {}
+
+OUTPUT_TYPES: Dict[str, TextIO] = {
+    'stdout': sys.stdout,
+    'stderr': sys.stderr
+}
 
 
-def get_variable(name, default=None):
+def get_variable(name: str, default: Optional[str] = None) -> Optional[str]:
     """Get Robot Framework variable.
 
     :param name:    Name of the variable
@@ -37,10 +45,34 @@ def get_variable(name, default=None):
         return _variables.get(name, default)
 
 
-class Variables(object):
+class Variables:
     """This class stores Robot Framework variables related to Report Portal."""
 
-    def __init__(self):
+    enabled: bool = ...
+    endpoint: Optional[str] = ...
+    launch_name: Optional[str] = ...
+    _pabot_pool_id: Optional[int] = ...
+    _pabot_used: Optional[str] = ...
+    project: Optional[str] = ...
+    api_key: Optional[str] = ...
+    attach_log: bool = ...
+    attach_report: bool = ...
+    attach_xunit: bool = ...
+    launch_attributes: list = ...
+    launch_id: Optional[str] = ...
+    launch_doc: Optional[str] = ...
+    log_batch_size: Optional[int] = ...
+    mode: Optional[str] = ...
+    pool_size: Optional[int] = ...
+    rerun: bool = ...
+    rerun_of: Optional[str] = ...
+    test_attributes: Optional[list] = ...
+    skipped_issue: bool = ...
+    log_batch_payload_size: int = ...
+    launch_uuid_print: bool
+    launch_uuid_print_output: TextIO
+
+    def __init__(self) -> None:
         """Initialize instance attributes."""
         self.endpoint = get_variable('RP_ENDPOINT')
         self.launch_name = get_variable('RP_LAUNCH')
@@ -70,8 +102,11 @@ class Variables(object):
         self.test_attributes = get_variable(
             'RP_TEST_ATTRIBUTES', default='').split()
         self.log_batch_payload_size = int(get_variable(
-            "RP_LOG_BATCH_PAYLOAD_SIZE",
+            'RP_LOG_BATCH_PAYLOAD_SIZE',
             default=str(MAX_LOG_BATCH_PAYLOAD_SIZE)))
+        self.launch_uuid_print = bool(strtobool(get_variable('RP_LAUNCH_UUID_PRINT', default='False')))
+        self.launch_uuid_print_output = OUTPUT_TYPES.get(
+            get_variable('RP_LAUNCH_UUID_PRINT_OUTPUT', default='stdout').lower(), OUTPUT_TYPES['stdout'])
 
         self.api_key = get_variable('RP_API_KEY')
         if not self.api_key:
@@ -106,7 +141,7 @@ class Variables(object):
             )
 
     @property
-    def pabot_pool_id(self):
+    def pabot_pool_id(self) -> int:
         """Get pool id for the current Robot Framework executor.
 
         :return: Pool id for the current Robot Framework executor
@@ -116,7 +151,7 @@ class Variables(object):
         return self._pabot_pool_id
 
     @property
-    def pabot_used(self):
+    def pabot_used(self) -> Optional[str]:
         """Get status of using pabot in test execution.
 
         :return: Cached value of the Pabotlib URI
@@ -126,7 +161,7 @@ class Variables(object):
         return self._pabot_used
 
     @property
-    def verify_ssl(self):
+    def verify_ssl(self) -> Union[bool, str]:
         """Get value of the verify_ssl parameter for the client."""
         verify_ssl = get_variable('RP_VERIFY_SSL', default='True')
         if path.exists(verify_ssl):