+++ tests/integration/test_no_keyword_message.py
@@ -15,7 +15,7 @@ limitations under the License
 
 import re
 from tests.helpers import utils
-from six.moves import mock
+from unittest import mock
 
 from tests import REPORT_PORTAL_SERVICE