+++ tests/unit/test_listener.py
@@ -14,7 +14,7 @@ limitations under the License
 """
 
 import pytest
-from six.moves import mock
+from unittest import mock
 
 from robotframework_reportportal.listener import listener
 from tests import REPORT_PORTAL_SERVICE