+++ tests/unit/test_logger.py
@@ -15,7 +15,7 @@ limitations under the License
 
 import pytest
 from delayed_assert import assert_expectations, expect
-from six.moves import mock
+from unittest import mock
 
 from robotframework_reportportal import logger